✨ feat(test/index.tsx): create a new TestPage component
This commit adds a new file `test/index.tsx` which contains the implementation of the TestPage component. The TestPage component is a functional component that fetches data from the specified API endpoint using the `useGetRequest` hook. It then sets the fetched data to the `data` state variable and renders a simple heading "Test Page".